requirements.txt
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
django-filter>=23.0
psycopg2-binary>=2.9.0
celery>=5.3.0
redis>=4.5.0
gunicorn>=21.2.0
